@@ -5,7 +5,7 @@
   "type": "module",
   "engines": {
     "node": ">=24.13.0",
-    "pnpm": "10.x"
+    "pnpm": ">=10.28.2"
   },
   "scripts": {
     "dev": "turbo dev",
@@ -22,13 +22,13 @@
     "infra:logs": "docker compose -f infra/docker-compose.dev.yml logs -f"
   },
   "devDependencies": {
-    "@eslint/js": "^9.19.0",
-    "@types/node": "^22.10.5",
-    "eslint": "^9.19.0",
-    "prettier": "^3.4.2",
+    "@eslint/js": "^9.39.2",
+    "@types/node": "^25.2.0",
+    "eslint": "^9.39.2",
+    "prettier": "^3.8.1",
     "turbo": "^2.8.3",
-    "typescript": "^5.9.0",
-    "typescript-eslint": "^8.24.0"
+    "typescript": "^5.9.3",
+    "typescript-eslint": "^8.54.0"
   },
-  "packageManager": "pnpm@10.0.0"
+  "packageManager": "pnpm@10.28.2"
 }
